Rename 'worker-harness' to 'worker'.
